requirements.txt
altair==5.5.0
GitPython==3.1.44
ipykernel==6.29.5
ipython==8.31.0
jedi==0.19.2
Jinja2==3.1.5
joblib==1.4.2
jsonschema==4.23.0
jupyter_client==8.6.3
jupyter_core==5.7.2
matplotlib==3.10.0
missingno==0.5.2
numpy==2.2.1
nvidia-nccl-cu12==2.24.3
pandas==2.2.3
patsy==1.0.1
pyarrow==18.1.0
pydeck==0.9.1
requests==2.32.3
scikit-learn==1.6.0
scipy==1.15.0
seaborn==0.13.2
statsmodels==0.14.4
streamlit==1.41.1
pytest
fastapi==0.115.12
alembic==1.15.2
email_validator==2.2.0
python-dotenv==1.1.0
python-jose==3.4.0
SQLAlchemy==2.0.40
sqlmodel==0.0.24
bcrypt==4.0.1
passlib==1.7.4
pydantic[email]
jose
uvicorn
python-multipart
